+++ Model/AdditionalAttributes/Provider/Doofinder.php
@@ -3,6 +3,7 @@
 namespace Doofinder\Feed\Model\AdditionalAttributes\Provider;
 
 use Doofinder\Feed\Model\AdditionalAttributes\AttributesProviderInterface;
+use Doofinder\Feed\Model\Config\Indexer\Attributes;
 use Doofinder\Feed\Helper\StoreConfig;
 
 /**
@@ -10,6 +11,11 @@ use Doofinder\Feed\Helper\StoreConfig;
  */
 class Doofinder implements AttributesProviderInterface
 {
+    /**
+     * @var Attributes
+     */
+    private $indexerAttributes;
+
     /**
      * @var StoreConfig
      */
@@ -22,10 +28,14 @@ class Doofinder implements AttributesProviderInterface
 
     /**
      * Doofinder constructor.
+     * @param Attributes $indexerAttributes
      * @param StoreConfig $storeConfig
      */
-    public function __construct(StoreConfig $storeConfig)
-    {
+    public function __construct(
+        Attributes $indexerAttributes,
+        StoreConfig $storeConfig
+    ) {
+        $this->indexerAttributes = $indexerAttributes;
         $this->storeConfig = $storeConfig;
     }
 
@@ -37,7 +47,9 @@ class Doofinder implements AttributesProviderInterface
     {
         if (!$this->attributes) {
             $this->attributes = array_keys(
-                $this->storeConfig->getDefaultDoofinderFields()
+                $this->indexerAttributes->get(
+                    $this->storeConfig->getCurrentStore()->getId()
+                )
             );
         }
         return $this->attributes;