+++ Model/Config/Source/Feed/Attributes.php
@@ -2,42 +2,31 @@
 
 namespace Doofinder\Feed\Model\Config\Source\Feed;
 
+use Magento\Framework\Option\ArrayInterface;
+
 /**
- * Attributes source
+ * Class Attributes
+ * The class responsible for providing options in system configuration
  */
-class Attributes implements \Magento\Framework\Option\ArrayInterface
+class Attributes implements ArrayInterface
 {
     /**
-     * Eav Model Config
-     *
-     * @var \Magento\Eav\Model\Config
+     * @var ArrayInterface[]
      */
-    private $eavConfig;
+    private $providers;
+
     /**
-     * Escaper
-     *
-     * @var \Magento\Framework\Escaper
-     */
-    private $escaper;
-    /**
-     * Array with Doofinder Options and Product Attributes.
-     *
      * @var array
      */
-    private $options;
+    private $options = [];
 
     /**
      * Attributes constructor.
-     *
-     * @param \Magento\Eav\Model\Config $eavConfig
-     * @param \Magento\Framework\Escaper $escaper
+     * @param array $providers
      */
-    public function __construct(
-        \Magento\Eav\Model\Config $eavConfig,
-        \Magento\Framework\Escaper $escaper
-    ) {
-        $this->eavConfig = $eavConfig;
-        $this->escaper = $escaper;
+    public function __construct(array $providers = [])
+    {
+        $this->providers = $providers;
     }
 
     /**
@@ -47,74 +36,16 @@ class Attributes implements \Magento\Framework\Option\ArrayInterface
      */
     public function toOptionArray()
     {
-        if (!$this->options) {
-            $this->options = array_merge(
-                $this->_getDoofinderDirectivesOptionArray(),
-                $this->_getCatalogProductAttributes()
-            );
+        if (empty($this->options)) {
+            foreach ($this->providers as $provider) {
+                if ($provider instanceof ArrayInterface) {
+                    $this->options = array_merge(
+                        $this->options,
+                        $provider->toOptionArray()
+                    );
+                }
+            }
         }
-
         return $this->options;
-    }
-
-    /**
-     * Getting all attributes into one array.
-     *
-     * @return array
-     */
-    public function getAllAttributes()
-    {
-        $doofinderOptions = $this->_getDoofinderDirectivesOptionArray();
-        $productAttributes = $this->_getCatalogProductAttributes();
-
-        foreach ($doofinderOptions as $code => $value) {
-            $value = $this->escaper->escapeJsQuote(($value));
-            $doofinderOptions[$code] = $value;
-        }
-
-        foreach ($productAttributes as $code => $value) {
-            $value = $this->escaper->escapeJsQuote(($value));
-            $productAttributes[$code] = $value;
-        }
-
-        return array_merge($doofinderOptions, $productAttributes);
-    }
-
-    /**
-     * Getting all product attributes.
-     *
-     * @return array
-     */
-    private function _getCatalogProductAttributes()
-    {
-        $attributes = [];
-
-        $productEntity = \Magento\Catalog\Model\Product::ENTITY;
-        $collection = $this->eavConfig->getEntityType($productEntity)->getAttributeCollection();
-
-        foreach ($collection as $attribute) {
-            $code = $attribute->getAttributeCode();
-            $label = $attribute->getFrontendLabel();
-
-            $attributes[$code] = 'Attribute: ' . $code . ($label ? ' (' . $label . ')' : '');
-        }
-
-        return $attributes;
-    }
-
-    /**
-     * Getting all doofinder options from directives in xml config.
-     *
-     * @return array
-     */
-    private function _getDoofinderDirectivesOptionArray()
-    {
-        return [
-            'df_id' => __('Doofinder: Product Id'),
-            'df_availability' => __('Doofinder: Product Availability'),
-            'df_currency' => __('Doofinder: Product Currency'),
-            'df_regular_price' => __('Doofinder: Product Regular Price'),
-            'df_sale_price' => __('Doofinder: Product Sale Price'),
-        ];
     }
 }